requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
exceptiongroup==1.2.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
yarl==1.9.3
